+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: aarias-d <aarias-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/04 16:58:37 by aarias-d          #+#    #+#             */
-/*   Updated: 2025/06/08 18:19:13 by aarias-d         ###   ########.fr       */
+/*   Updated: 2025/06/08 19:35:10 by aarias-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 int		ft_printf(char const *, ...);
 int		ft_putstr_fd(char *s, int fd);
-int		ft_putnbr_fd(int n, int fd);
+int		ft_putnbr_fd(long n, int fd);
 int		ft_putchar_fd(char c, int fd);
 int		ft_toupper(int c);
 int		ft_puthexadecimal(unsigned int  num, char x);
